+++ .env
@@ -1,1 +1,1 @@
-DB_HOST=mongodb+srv://natalia_goryanina:2ty37B3LjxgpDQN@cluster0.xhcz3.mongodb.net/Cluster0?retryWrites=true&w=majority
+DB_HOST=mongodb+srv://natalia_goryanina:2ty37B3LjxgpDQN@cluster0.xhcz3.mongodb.net/db-contacts?retryWrites=true&w=majority
