+++ nodemon.json
@@ -1,3 +1,3 @@
 {
-  "ignore": ["node_modules", "model/contacts.json"]
+  "ignore": ["node_modules", "service/model/contacts.json"]
 }
